Former-commit-id: 1004a4f1ab24e173ed67a786bd8d6bdff4364b59

+++ modules/image/producer/image_scroll_producer.cpp
@@ -401,7 +401,7 @@ struct image_scroll_producer : public core::frame_producer_base
 			auto diff = *end_time_ - now;
 			auto seconds = diff.total_seconds();
 
-			set_speed(-speed_from_duration(seconds));
+			set_speed(-speed_from_duration(static_cast<double>(seconds)));
 			end_time_ = boost::none;
 		}
 		else
